Added a very basic physics system.

+++ src/components/position.h
@@ -9,9 +9,9 @@ namespace components
 
 struct Position
 {
-    Position(int x, int y) : x(x), y(y) {}
+    Position(float x, float y) : x(x), y(y) {}
 
-    int x, y;
+    float x, y;
 };
 
 } // end components
